+++ src/components/manage/Display/Display.jsx
@@ -5,7 +5,6 @@ import { compose } from 'redux';
 import { injectLazyLibs } from '@plone/volto/helpers/Loadable/Loadable';
 
 import { getSchema, updateContent, getContent } from '@plone/volto/actions';
-import layouts from '@plone/volto/constants/Layouts';
 import { getLayoutFieldname } from '@plone/volto/helpers';
 import { FormFieldWrapper, Icon } from '@plone/volto/components';
 import { defineMessages, injectIntl } from 'react-intl';
@@ -139,7 +138,12 @@ class DisplaySelect extends Component {
   state = {
     selectedOption: {
       value: this.props.layout,
-      label: layouts[this.props.layout] || this.props.layout,
+      label:
+        this.props.intl.formatMessage({
+          id: config.views.layoutViewsNamesMapping?.[this.props.layout],
+          defaultMessage:
+            config.views.layoutViewsNamesMapping?.[this.props.layout],
+        }) || this.props.layout,
     },
   };
 
@@ -191,6 +195,7 @@ class DisplaySelect extends Component {
   render() {
     const { selectedOption } = this.state;
     const Select = this.props.reactSelect.default;
+    const layoutsNames = config.views.layoutViewsNamesMapping;
 
     return (
       <FormFieldWrapper
@@ -210,7 +215,11 @@ class DisplaySelect extends Component {
             )
             .map((item) => ({
               value: item,
-              label: layouts[item] || item,
+              label:
+                this.props.intl.formatMessage({
+                  id: layoutsNames[item],
+                  defaultMessage: layoutsNames[item],
+                }) || item,
             }))}
           styles={customSelectStyles}
           theme={selectTheme}